pyproject.toml
[tool.poetry]
name = "photo_app"
version = "0.1.0"
description = "Simple photo management app"
authors = ["Yann Rolland"]
packages = [ { include = "src" } ]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
PySide6 = "^6.6.0"
Pillow = "^10.0.0"
piexif = "^1.1.3"
opencv-python = "^4.9.0.80"
imagehash = "^4.3.2"
scikit-learn = "^1.5.0"
pyqtdarktheme2 = "^2.1.2"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
xgboost = "^3.1.1"
# Web API dependencies
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.17"
aiosqlite = "^0.20.0"
greenlet = "^3.2.4"
torch = "^2.9.0"
torchvision = "^0.24.0"
tqdm = "^4.67.1"
optuna = "^4.6.0"
datasets = "^4.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-cov = "^4.1"
ruff = "^0.6"
mypy = "^1.13"
types-Pillow = "^10.2"
requests = "^2.32.5"
torch = "^2.9.0"
torchvision = "^0.24.0"
joblib = "^1.5.2"
numpy = "^2.3.4"
pytest-timeout = "^2.4.0"
pytest-asyncio = "^1.3.0"
isort = "^7.0.0"
pyinstrument = "^5.1.1"
lightgbm = "^4.6.0"
catboost = "^1.2.8"
py-spy = "^0.4.1"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
